backend/requirements.txt
asgiref==3.8.1
astroid==3.3.10
attrs==25.3.0
boltons==21.0.0
boto3==1.38.40
botocore==1.38.40
bracex==2.5.post1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.7
colorama==0.4.6
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.4.0
Django==5.2.3
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
face==24.0.0
glom==22.1.0
googleapis-common-protos==1.70.0
idna==3.10
importlib_metadata==7.1.0
isort==6.0.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==25.0
pathspec==0.12.1
pbr==6.1.1
peewee==3.18.1
platformdirs==4.3.8
protobuf==4.25.8
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
pycodestyle==2.13.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rich==13.5.3
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.13.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stevedore==5.4.1
tomli==2.0.2
tomlkit==0.13.3
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcmatch==8.5.2
wrapt==1.17.2
zipp==3.23.0
